requirements.txt
anthropic
python-dotenv
yfinance
requests
beautifulsoup4
pandas
streamlit